added markdown preview for articles

+++ src/components/SinglePost.js
@@ -3,7 +3,7 @@ import React, { useContext, useEffect, useState } from 'react'
 import { UserContext } from './userContext'
 import { Remarkable } from 'remarkable';
 import { Button, Container } from 'react-bootstrap'
-import {Link, useHistory, useParams } from 'react-router-dom'
+import { Link, useHistory, useParams } from 'react-router-dom'
 import postService from '../services/postService';
 
 
@@ -24,6 +24,7 @@ const SinglePost = () => {
   }
 
   const handleDeletion = async () => {
+    if (!window.confirm(`Удалить статью "${post.title}"?`)) return
     await postService.deletePost(post.id)
     history.push('/')
   }
@@ -31,6 +32,7 @@ const SinglePost = () => {
     width: "55%",
   };
   if (post === null) return (<></>)
+  
   return (
     <Container>
       <br />
@@ -38,13 +40,14 @@ const SinglePost = () => {
         <h1>{post.title}</h1>
         {(user && (user.id === post.author.id || user.role === 'admin')) &&
           <div className="ml-auto">
-            <Button variant="outline-secondary">
+            {post &&
               <Link to={'/edit/' + post?.id}>
-                <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" className="bi bi-pencil" viewBox="0 0 16 16">
-                  <path d="M12.146.146a.5.5 0 0 1 .708 0l3 3a.5.5 0 0 1 0 .708l-10 10a.5.5 0 0 1-.168.11l-5 2a.5.5 0 0 1-.65-.65l2-5a.5.5 0 0 1 .11-.168l10-10zM11.207 2.5 13.5 4.793 14.793 3.5 12.5 1.207 11.207 2.5zm1.586 3L10.5 3.207 4 9.707V10h.5a.5.5 0 0 1 .5.5v.5h.5a.5.5 0 0 1 .5.5v.5h.293l6.5-6.5zm-9.761 5.175-.106.106-1.528 3.821 3.821-1.528.106-.106A.5.5 0 0 1 5 12.5V12h-.5a.5.5 0 0 1-.5-.5V11h-.5a.5.5 0 0 1-.468-.325z" />
-                </svg>
-              </Link>
-            </Button>
+                <Button variant="outline-secondary">
+                  <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" className="bi bi-pencil" viewBox="0 0 16 16">
+                    <path d="M12.146.146a.5.5 0 0 1 .708 0l3 3a.5.5 0 0 1 0 .708l-10 10a.5.5 0 0 1-.168.11l-5 2a.5.5 0 0 1-.65-.65l2-5a.5.5 0 0 1 .11-.168l10-10zM11.207 2.5 13.5 4.793 14.793 3.5 12.5 1.207 11.207 2.5zm1.586 3L10.5 3.207 4 9.707V10h.5a.5.5 0 0 1 .5.5v.5h.5a.5.5 0 0 1 .5.5v.5h.293l6.5-6.5zm-9.761 5.175-.106.106-1.528 3.821 3.821-1.528.106-.106A.5.5 0 0 1 5 12.5V12h-.5a.5.5 0 0 1-.5-.5V11h-.5a.5.5 0 0 1-.468-.325z" />
+                  </svg>
+                </Button>
+              </Link>}
             <Button variant="outline-danger" onClick={handleDeletion}>
               <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" className="bi bi-trash" viewBox="0 0 16 16">
                 <path d="M5.5 5.5A.5.5 0 0 1 6 6v6a.5.5 0 0 1-1 0V6a.5.5 0 0 1 .5-.5zm2.5 0a.5.5 0 0 1 .5.5v6a.5.5 0 0 1-1 0V6a.5.5 0 0 1 .5-.5zm3 .5a.5.5 0 0 0-1 0v6a.5.5 0 0 0 1 0V6z" />
@@ -56,7 +59,7 @@ const SinglePost = () => {
       <div><em>{post.author.username}</em></div>
       {/*<div>{post.content}</div>*/}
       <div className="content" dangerouslySetInnerHTML={getRawMarkup()}></div>
-    </Container>
+    </Container >
   )
 }
